Always select a valid move, even if it's losing.
Before this change the player would only select an inferred move
(which is optimal if there is one, because it doesn't change the
solution set). After this change, the player will select a losing
move if the set of inferred moves is empty, with preference given
to moves that maximize the total number of remaining solutins.