Refactor SelectMoveFromSolutions() to return outcome.
This allows the player to detect when a state switches from "winning" to
"losing", which would imply there was a bug in the analysis.

I also removed the output to std::cerr from analysis.cc, so that I can customize
the output in solver and player independently. (I'm planning to introduce
structured logging in the player later.)