Refactor time accounting to (hopefully) fix inaccurate time tracking on CodeCup systems.
The time I log on the CodeCup server is much higher than the time used
according to the CodeCup server. I think the reason was that the referee
suspends my player immediately after writing its output, so I should
pause the timer before writing the output, instead of blocking before
reading the opponent's move, which is what I did before.