Don't sort moves fully, but extract from a heap instead.
This seems to improve performance by around 20%. The reason this is more
performant is that in many cases we only need to evaluate part of the ranked
move list to find a winning move, so instead of spending O(N log N) where N is
the number of available moves, we spend O(N + M log N) time, where M is the
number of moves we actually look at. This is faster when M is significantly
smaller than N on average.