Don't try to enumerate solutions until at least 8 clues have been placed.
This saves a little time on enumerating solutions very early in the game, when
it doesn't tell us much. At the same time, increase the maximum number of
solutions to enumerate to 400,000, to make sure we don't reduce the solution
count too much early in the midgame (before full analysis is possible).